+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Test and beta releases will have experimental functions enabled by default.
 * Less restrictive database file permissions, always set on boot
 * Support for multi SIM IPC restrictions (Huawei and maybe others)
 * Added *act on category* check box to restricting on demand prompt, thanks @[jpeg729](https://github.com/jpeg729)
+* Added application icon to restricting on demand prompt
 * On demand restricting default enabled (there is a new main setting to disable it)
 * If LBE Security Master is enabled, only applications can be restricted ([issue](https://github.com/M66B/XPrivacy/issues/1231))
 * Fixed randomize on access not sticking for latitude and longitude